fix: use .toLong for slotToTime to support Scala.js BigInt
On Scala.js, slotToTime returns Double and BigInt(Double) doesn't
exist. Also remove unused imports in StateLocator.scala.

+++ scalus-utxo-cell/shared/src/test/scala/scalus/utxocells/AuctionCellSpec.scala
@@ -150,7 +150,7 @@ class AuctionCellSpec extends AnyFunSuite {
 
     // Auction end time: slot 100 in mainnet config
     private val auctionEndTime: PosixTime = BigInt(
-      testEnv.slotConfig.slotToTime(100)
+      testEnv.slotConfig.slotToTime(100).toLong
     )
 
     private def createEmulator(): Emulator = Emulator(
